Add colGap on Row, and rename its extension to tsx
